devstackディレクトリーもif文で取り扱うように修正

+++ devstack.sh
@@ -1,20 +1,23 @@
-#!/bin/bash -x
+#!/bin/bash
 
-rm -rf devstack
-git clone -b stable/kilo https://git.openstack.org/openstack-dev/devstack
+if [ -d devstack ];
+then
+  echo "devstack dir is found. Clone the New Sources."
+  rm -rf devstack
+  git clone -b stable/kilo https://git.openstack.org/openstack-dev/devstack
+else
+  echo "devstack dir is not found. Clone the Sources."
+  git clone -b stable/kilo https://git.openstack.org/openstack-dev/devstack
+fi
 
 if [ -h devstack/local.conf ];
 then
    echo "local.conf is found."
 else
    ln -s ~/devstack-memo/local.conf devstack/local.conf
-   echo "ln -s is done."
+   echo "ln -s local.conf is done."
 fi
 
 echo "Run the stack.sh? Cancel:(ctrl+c)"
 read confstack1
 bash devstack/stack.sh
-
-echo "Configure br-ex? Cancel:(ctrl+c)"
-read confstack2
-sudo ifconfig br-ex 0.0.0.0 && sudo ovs-vsctl add-port br-ex eth0
